[sequential SfM] add image scoring strategy for next best view and initial image pair selection
Image score is based on a pyramid grid. The goal of the pyramid grid is
to promote a good repartition in the image without expensive
computation.

* Next best view:
** As we have a fixed number of grid cells, we use a constant threshold
on the score to choose the next best views instead of a percentage of
the best score. The goal is to avoid the selection of many images in
critical steps of the reconstruction (when the best score is low, many
images are selected). The current drawback is that at the end of the
reconstruction all the bad images generates a lot of BA, which cost a
lot because we are at the end of the reconstruction.
** We use a max threshold on the number of images to add to avoid adding
to much data between 2 BA.
** If one intrinsic is used for the first time, we ensure that a global
BA is performed just after.
** If one intrinsic is used for the first time, we refine the intrinsics
in RefinePose before adding new 3D points.
** Add a min number of points.
** We precompute a cell index in the pyramid grid for each track in each
input view for efficiency.
** Perform a full BA for the 30 first cameras. It's critical at the
beginning and it's also cheap as we have limited data.

* Initial image pair:
** The score of each image pair is based on the combination of the image
score and the angle (not only the angle).
imagePairScore = min(scoreImageA, scoreImageB) * angle.
** increase the max_iteration_count of the robustRelativePose to avoid
too much randomness