[build] Cmake enhancement #798
- Linear programming related:

- Remove the code that was using MOSEK for Linear Programming:
    - I kept it as an option for the LP solver but since no one show
interest about using it. I prefer to remove it to keep the code
lightweight (less code to maintain).

- Move the OSI_X_SolverWrapper implementation to the
openMVG_linearProgramming library

- Due to the previous point it is now possible to hide the OSI CLP
header and code to a cpp file and so to use PRIVATE include directories
related to OSI CLP. It means that less include directories and link
libraries will be used when OpenMVG is used as a third party library.