[matching] Simpler hamming_distance design #1064
openMVG/matching/metric_hamming.hpp presents too many system
specialization.

Her we revert the Hamming distance class to a simpler design that rely
on using bitset::count in order to ease the maintenance in the future.

Here some notes:

- Since the appropriate compiler flag are set, the cpu instruction
popcnt will be used and so the code will be as fast as the hand tuned
called (at least for GCC and clang).