[build/install] Fix Eigen install path issue. #1335 (#1117, #1118)
Since there is a bug in Eigen cmake file with the usage of
`INCLUDE_INSTALL_DIR` and `EIGEN_INCLUDE_INSTALL_DIR` we force here the
usage of `EIGEN_INCLUDE_INSTALL_DIR`. It avoids the definition of
`${CMAKE_INSTALL_INCLUDEDIR}/eigen3` by default that will go to
`/usr/include` even we use a local specified directory.