Small typo in documentation (#1586)
Fix typo: ou -> or
I also added the world non-coplanar since correspondences are likely and expected to be not on the same plane, otherwise, it creates degeneracy which makes it impossible to correctly calculate the F matrix.
Minimum number correspondences required is 7. But since it was already 8 ou 7 I changed it to 8 or 7.