[Trifocal] fixed retry logic. No
Noticed that trifocal_test for solve never uses solve retries based on
the return value of minus::solve(). But it really the solve may fail
some 5% of the time only because the ground truth solution has not been
found. This merits closer exploration of the minus_io_14a<P, F>::
probe_all_solutions function and if it really did not find the ground
truth and exactly why. A study of the fail cases is needed.
However, by rerunning minus::solve in case the ground-truth has not been
found, it will find it most of the time. So a single re-run might give
99.99% success rate. So a single rerun will give aditional 212 roots,
so if we always search among 212*2 roots we will be 99% safe. However,
this likely is not needed since the solver runs inside RANSAC.