Remove using namespace std and add std:: qualifications where needed (#2092)
`using namespace std;` is generally an antipattern as it imports a huge
number of common names into the global namespace. We already qualify
names with std:: in majority of cases and `using namespace std` is not
actually needed most of the time. This commit addresses a small number
of remaining cases where std:: is missing and removes all `using
namespace std`.