[C++] prefer push_back(std::move(X))
Here what is preferred:
- Use push_back by default.
- Use emplace_back where it is semantically significant to your
algorithm (such as when the element type’s move-constructor is absent or
has been benchmarked as expensive).

It was unfortunate but we have used emplace_back(X) and thought it would
avoid a copy of the object. We are fixing this here by using
push_back(std::move(X))